pyproject.toml
[tool.poetry]
name = "web3-playground"
version = "0.1.0"
description = ""
authors = ["sgerodes <sgerodes@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
eth-account = "^0.13.0"
eciespy = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
